Descent v1: Player Movement
The game will eventually be a first person maze exploration game, with the goal being to navigate a series of mazes without dying. This is a rough prototype working out 1. how to change player position, and 2. how to represent that graphically for the player. The player is placed in a grid (size controlled by rmSize) and by inputting directional commands the player (represented by '#') will move through the grid. Need to work: how to define inner walls of maze; proper player control (forward and turning), how to hide maze walls until discovered by player.