Descent v2 - Wall Detection (a real maze!!!)
While the player was prevented from leaving the maze, there were no internal walls to restrict movement. Now, internal walls are represented in two matrices (wallH and wallV). When the player moves, the game checks if there is a wall based on the players position and direction. Internal walls are also represented graphically.
Need to add -  hide walls not seen by player, how to adapt system to larger room sizes (so as not to clutter screen), maze randomization (?)
Scrapped ideas - first person: this is too difficult to implement, due to the complexity of the necessary code and my lack of ASCII art skills. The UI will only consist of the maze map, with potentially a separate menu for combat(?)